+++ lib/index.js
@@ -1,6 +1,4 @@
-var post = require('./post');
-var page = require('./page');
-var build = require('./build');
+'use strict';
 
 module.exports.post = require('./post');
 module.exports.page = require('./page');
