+++ lib/page.js
@@ -5,32 +5,19 @@ var fs = require('fs');
 var mkdirp = require('mkdirp');
 var moment = require('moment');
 
+var config = require('./config.json');
 var utils = require('./utils');
-var build = require('./build');
 
-var t = title || null,
-    s = slug || null;
+var title, slug = "Page title example";
 
 function page(title) {
     console.log('Console Message: Building Page');
-
-    // If the output is either -t or --title  then we store this to title
     if (title) {
-
-        var title = argv.title || argv.t;
-        var slug = (argv.slug || argv.s) ? utils.hyphenateString(argv.slug || argv.s) : utils.hyphenateString(title.toString());
-        
-        // Filename has the value of the file path of the page in question
-        var filename = build.srcDir.pages + slug + '.md';
-
-        // fs.access determines if a file exists, regardless of rwx permissions
-        fs.access(filename, fs.F_OK, function (err) {
-
-            // Since this exists, then we can log to the console that this file(in question) exists
+        var filename = config.pages + slug + '.md';
+        fs.access(filename, fs.F_OK, function(err) {
             if (err) {
                 return 'Task aborted: File ' + filename + ' already exists.';
             }
-
             var meta =
                 '---\r\n' +                                 // --- CR/NL regex
                 'date: ' + moment().format() + '\r\n' +     // Date format, CR/NL regex
@@ -43,34 +30,18 @@ function page(title) {
                 '---\r\n' +                                 // --- CR/NL regex
                 ' <Insert your content here> \r\n ';        // Content goes here CR/NL regex
 
-            // We're creating a directory corresponding to the paths.page string (filepath).             
-            mkdirp(build.srcDir.pages, function (err) {
+            mkdirp(config.pages, function(err) {
+                if (err) return err;
 
-                // If error, return error.
-                if (err) {
-                    return err;
-                }
-
-                // FileSystem will now write the file according to filename/meta
-                fs.writeFile(filename, meta, function (err) {
-
-                    // Error handler
-                    if (err) {
-                        return err;
-                    }
-
-                    // Log to the console the success of the page generated
+                fs.writeFile(filename, meta, function(err) {
+                    if (err) return err;
                     console.log('Task success - Your new page was successfully created: ', filename);
                 });
             });
         });
-
     } else {
-
-        // Otherwise we log the error out to console if user didn't input -t or --title
         return 'Error: You must enter a title for your page. Use --title "Page title" or -t "Page title"';
     }
-
 }
 
 module.exports = page;
